# server/src/project/project.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import { ProjectService } from './project.service';
import { CreateProjectDto } from './dto/create-project.dto';
import { UpdateProjectDto } from './dto/update-project.dto';
import { ApiBody } from '@nestjs/swagger';

@Controller('project')
export class ProjectController {
  constructor(private readonly projectService: ProjectService) {}

  @ApiBody({ type: CreateProjectDto })
  @Post()
  create(@Body() createProjectDto: CreateProjectDto) {
    return this.projectService.create(createProjectDto);
  }

  @Get()
  findAll() {
    return this.projectService.findAll();
  }

  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.projectService.findOne(id);
  }

  @Get(':id/issues')
  findIssues(@Param('id') id: string) {
    return this.projectService.findIssues(id);
  }

  @Get(':id/members')
  findMembers(@Param('id') id: string) {
    return this.projectService.findMembers(id);
  }

  @ApiBody({ type: UpdateProjectDto })
  @Patch(':id')
  update(@Param('id') id: string, @Body() updateProjectDto: UpdateProjectDto) {
    return this.projectService.update(id, updateProjectDto);
  }

  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.projectService.remove(id);
  }

  @Get(':id/issue-types')
  findIssueTypes(@Param('id') id: string) {
    return this.projectService.findIssueTypes(id);
  }

  @Get(':projectId/issue-type/:issueTypeId/fields')
  findFieldsByIssueType(
    @Param('issueTypeId') issueTypeId: string,
    @Param('projectId') projectId: string,
  ) {
    console.log(issueTypeId, projectId);
    return this.projectService.findFieldsByIssueType(issueTypeId, projectId);
  }
}